Complete CUINamedColor, CUINamedLookup

+++ CoreUI/include/CUIColorName.h
@@ -11,7 +11,7 @@
 #import <CoreFoundation/CoreFoundation.h>
 
 // TODO: 29 cases
-typedef CF_CLOSED_ENUM(uint64_t, CUIColorName) {
+typedef CF_CLOSED_ENUM(NSUInteger, CUIColorName) {
     CUIColorNameRed = 0,
     CUIColorNameOrange = 1,
     CUIColorNameYellow = 2,
