+++ CoreUI/include/CUIDisplayGamut.h
@@ -10,7 +10,7 @@
 
 #import <CoreFoundation/CoreFoundation.h>
 
-typedef CF_CLOSED_ENUM(uint64_t, CUIDisplayGamut) {
+typedef CF_CLOSED_ENUM(NSUInteger, CUIDisplayGamut) {
     CUIDisplayGamutSRGB = 0,
     CUIDisplayGamutP3 = 1,
 };
